+++ src/main/java/com/prgrms/be/intermark/domain/performance/model/Genre.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.intermark.domain.performance;
+package com.prgrms.be.intermark.domain.performance.model;
 
 public enum Genre {
     ROMANCE, COMEDY, HORROR, SF, DRAMA
