+++ src/main/java/com/prgrms/be/intermark/domain/performance/model/Performance.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.intermark.domain.performance;
+package com.prgrms.be.intermark.domain.performance.model;
 
 import com.prgrms.be.intermark.domain.casting.Casting;
 import com.prgrms.be.intermark.domain.performance_stadium.PerformanceStadium;
