+++ src/main/java/com/prgrms/be/intermark/domain/performance/model/PerformanceRating.java
@@ -1,4 +1,4 @@
-package com.prgrms.be.intermark.domain.performance;
+package com.prgrms.be.intermark.domain.performance.model;
 
 public enum PerformanceRating {
     ALL, ADULT, SEVEN, FIFTEEN
